+++ .gitignore
@@ -1,1 +1,1 @@
-creds.py
+backend/creds.py
